requirements.txt
google-genai==1.27.0
pandas==2.3.1
pydantic
python-dotenv==1.1.1
streamlit==1.47.1
tqdm==4.67.1